modules/sep-auth/src/main/java/com/silenteight/sep/auth/authorization/RoleAccessor.java
package com.silenteight.sep.auth.authorization;

import lombok.RequiredArgsConstructor;
import lombok.extern.slf4j.Slf4j;

import org.jetbrains.annotations.NotNull;
import org.springframework.security.access.AccessDeniedException;
import org.springframework.security.core.Authentication;
import org.springframework.security.core.GrantedAuthority;
import org.springframework.security.core.context.SecurityContextHolder;

import java.util.Set;
import java.util.stream.Collectors;

@Slf4j
@RequiredArgsConstructor
public class RoleAccessor {

  public Set<String> getRoles() {
    Authentication authentication = getAuthentication();
    return authentication
        .getAuthorities()
        .stream()
        .map(GrantedAuthority::getAuthority)
        .collect(Collectors.toSet());
  }

  @NotNull
  private static Authentication getAuthentication() {
    Authentication authentication = SecurityContextHolder.getContext().getAuthentication();

    if (authentication == null) {
      throw new AccessDeniedException("There is no authenticated principal");
    }
    return authentication;
  }
}
